Fix CalculatorTests

+++ TextCalculator.Tests/CalculatorTests.cs
@@ -8,8 +8,8 @@ namespace TextCalculator.Tests
     public class CalculatorTests
     {
         [Theory]
-        [InlineData("1+(2+3)*(4+5)", 101)]
-        public void Calculate_Expression_ReturnTrue(string expression, int expectedResult)
+        [InlineData("1+(2+3)*(4+5)", 101d)]
+        public void Calculate_Expression_ReturnTrue(string expression, double expectedResult)
         {
             // Arrange
             var calculator = new Calculator();
